user insert pp changed to get uin from session instead of input form

+++ user_func/user_program_progress/delete_pp/delete_class_data.php
@@ -14,18 +14,20 @@
 		die("Connection Failed : " . $conn->connect_error);
 	} else {
 
-        // delete certification record
-		$delete_class = "DELETE FROM Class_Enrollment WHERE CE_Num = '" . $input_CE_Num . "';";
-		$result_class = $conn->query($delete_class);
-        echo $delete_class;
+        $check_query = "SELECT * FROM Class_Enrollment WHERE CE_Num = '" . $input_CE_Num . "';";
+        $check_result = $conn->query($check_query);
 
-        if (!$result_class) {
-            echo "Could not delete class<br>";
-        } else {
+        if ($check_result->num_rows >= 1) {
+            // delete class record
+            $delete_class = "DELETE FROM Class_Enrollment WHERE CE_Num = '" . $input_CE_Num . "';";
+            $result_class = $conn->query($delete_class);
             echo "Deleted class successfully!<br>";
-            header("Location: ../program_progress_page.php");
+            echo "<a href='../program_progress_page.php'>Go back to program progress page</a>";
         }
-        
+        else{
+            echo "That class enrollment does not exist could not delete<br>";
+            echo "<a href='../program_progress_page.php'>Go back to program progress page</a>";
+        }
         
 	}
 ?>
