+++ user_func/user_program_progress/delete_pp/delete_internship_data.php
@@ -14,17 +14,20 @@
 		die("Connection Failed : " . $conn->connect_error);
 	} else {
 
-        // delete certification record
-		$delete_internship = "DELETE FROM Intern_App WHERE IA_Num = '" . $input_IA_Num . "';";
-		$result_internship = $conn->query($delete_internship);
+        $check_query = "SELECT * FROM Intern_App WHERE IA_Num = '" . $input_IA_Num . "';";
+        $check_result = $conn->query($check_query);
 
-        if (!$result_internship) {
-            echo "Could not delete internship<br>";
-        } else {
+
+        if ($check_result->num_rows >= 1) {
+            // delete certification record
+            $delete_internship = "DELETE FROM Intern_App WHERE IA_Num = '" . $input_IA_Num . "';";
+            $result_internship = $conn->query($delete_internship);
             echo "Deleted internship successfully!<br>"; 
-            header("Location: ../program_progress_page.php");
+            echo "<a href='../program_progress_page.php'>Go back to program progress page</a>";
         }
-        
-        
+        else {
+            echo "That internship application does not exist could not delete<br>";
+            echo "<a href='../program_progress_page.php'>Go back to program progress page</a>";
+        }
 	}
 ?>
